requirements.txt
Flask
twilio
tungsten
gunicorn
google-cloud-vision
google-cloud-storage
google-cloud-datastore
